requirements.txt
click
pyenchant